cmam_backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
joblib==1.5.3
scikit-learn==1.8.0
numpy==2.4.2
pandas==3.0.0
